feat: get cli working and fix typing
