+++ crates/goose-cli/Cargo.toml
@@ -13,7 +13,9 @@ path = "src/main.rs"
 
 [dependencies]
 goose = { path = "../goose" }
+developer = { path = "../developer" }
 mcp-client = { path = "../mcp-client" }
+mcp-server = { path = "../mcp-server" }
 mcp-core = { path = "../mcp-core" }
 clap = { version = "4.4", features = ["derive"] }
 cliclack = "0.3.5"
@@ -48,6 +50,7 @@ async-trait = "0.1"
 rustyline = "15.0.0"
 rust_decimal = "1.36.0"
 rust_decimal_macros = "1.36.0"
+tracing = "0.1.41"
 
 [dev-dependencies]
 tempfile = "3"
