+++ crates/goose-cli/src/commands/mod.rs
@@ -1,4 +1,5 @@
 pub mod agent_version;
 pub mod configure;
+pub mod server;
 pub mod session;
 pub mod version;
