+++ crates/goose/src/agents/mod.rs
@@ -5,7 +5,7 @@ mod factory;
 mod system;
 
 pub use agent::Agent;
+pub use capabilities::Capabilities;
 pub use default::DefaultAgent;
 pub use factory::{register_agent, AgentFactory};
-pub use capabilities::Capabilities;
 pub use system::SystemConfig;
