+++ crates/goose/src/providers/openai_utils.rs
@@ -1,4 +1,3 @@
-use mcp_core::ToolError;
 use crate::message::{Message, MessageContent};
 use crate::providers::base::Usage;
 use crate::providers::configs::ModelConfig;
@@ -7,6 +6,7 @@ use crate::providers::utils::{
     ImageFormat,
 };
 use anyhow::{anyhow, Error};
+use mcp_core::ToolError;
 use mcp_core::{Content, Role, Tool, ToolCall};
 use serde_json::{json, Value};
